Fix task arrangement and add checkmark to tasks
Removed unnecessary commented out code. Targeted the taskFormContainer
and the taskBlock container to remove and add to the DOM as needed to
get the list to properly display the task form and option to add new
tasks in the correct order and position. Added the checkmark svg when
user clicks the empty circle to visually show the task is finished.