Update delete functionality
Used event listeners to only show the delete task option when user
hovers over the task and to delete the fullTaskContainer div element
when clicking the delete icon. Targeted the container holding the delete
svg so only the one instance of the delete svg associated with the task
is affected.