Fix UI issues regarding default tasks
Update CSS styling of the dummy input element to differentiate it from
the form input elements. Applied styling to default tasks to highlight
the task that user is hovering above. Moved the piece of code appending
the taskPriorityBox inside the if else statements related to the task
priorities to improve appearance of tasks regardless of whether the user
selects a priority level. Created a clone of the edit svg to prevent the
svg from disppearing whenever the submit button was pressed. Made the
edit svg of tasks visible only when user hovers over the task. Updated
the styling of date popups to make them more visible and readable.