Fix user interface label selection issues
Used requerying to get newly created label options on the default list
form to create new labels and apply them to the list when selected.
Created a labelDuplicate variable and updated condtions to prevent
duplicates of existing labels. Restyled the "Create New Label" form to
reset style when the cancel button is pressed.