Fix task counting and statement issue and styling
Created an updateStatement function to handle updating the list's task
statement separate from the functions that handle opening and closing
the list. Fixed alignment issues with the dummy task input element
relative to tasks. Prevented task interactive when the list is closed
so users cannot alter the list unless opened.